Copyright page, page size.
Added a separate copyright page. This allows everything to be controlled from the main tex file. Added information about the GNU GPL (as well as a copy of the GNU GPL).

Changed page size so that it is more in line with a normal sized book. Footnotes are almost a problem. They really need to be renumbered by verse.